fix: 修复无法多开的问题

+++ src/BedrockBoot/Models/Pack/Game/Mods/ModsCore.cs
@@ -106,19 +106,22 @@ public class ModsCore
         {
             File.Copy(f.File, Path.Combine(preLoadPath, Path.GetFileName(f.File)));
         });
-        
-        string file_path = body;
-        if (File.Exists(file_path))
+
+        if (!FileCheck.IsFileLocked(body))
         {
-            using (PeFile peFile = new PeFile(File.Open(file_path, FileMode.OpenOrCreate, FileAccess.ReadWrite)))
-            using (var stream = AssetLoader.Open(new Uri("avares://BedrockBoot/Assets/PreloadCpp.dll")))
-            using (var memoryStream = new MemoryStream())
+            string file_path = body;
+            if (File.Exists(file_path))
             {
-                stream.CopyTo(memoryStream);
-                peFile.AddImport("PreloadCpp.dll", "Load");
-                peFile.Flush();
-                var fullPath = Path.Combine(VersionInfo.VersionPath, "PreloadCpp.dll");
-                File.WriteAllBytes(fullPath, memoryStream.ToArray());
+                using (PeFile peFile = new PeFile(File.Open(file_path, FileMode.OpenOrCreate, FileAccess.ReadWrite)))
+                using (var stream = AssetLoader.Open(new Uri("avares://BedrockBoot/Assets/PreloadCpp.dll")))
+                using (var memoryStream = new MemoryStream())
+                {
+                    stream.CopyTo(memoryStream);
+                    peFile.AddImport("PreloadCpp.dll", "Load");
+                    peFile.Flush();
+                    var fullPath = Path.Combine(VersionInfo.VersionPath, "PreloadCpp.dll");
+                    File.WriteAllBytes(fullPath, memoryStream.ToArray());
+                }
             }
         }
     }
